fix: merge into a single component
