+++ web-admin/src/features/embeds/TopNavigationBarEmbed.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-  import BreadcrumbItemEmbed from "@rilldata/web-common/components/navigation/breadcrumbs/BreadcrumbItemEmbed.svelte";
+  import BreadcrumbItem from "@rilldata/web-common/components/navigation/breadcrumbs/BreadcrumbItem.svelte";
   import { useValidDashboards } from "@rilldata/web-common/features/dashboards/selectors";
   import { ResourceKind } from "@rilldata/web-common/features/entity-management/resource-selectors";
   import type {
@@ -69,7 +69,7 @@
           </div>
         {/if}
         {#if currentResource}
-          <BreadcrumbItemEmbed
+          <BreadcrumbItem
             options={breadcrumbOptions}
             current={currentResourceName}
             onSelect={onSelectResource}
